backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-CORS==4.0.0
python-dotenv==1.0.0
firebase-admin==6.2.0
requests==2.31.0
psycopg2-binary==2.9.7
reportlab==4.0.4
gunicorn==21.2.0
Pillow==10.0.0 